Fix gcc compiler warnings reported by Sergei   modified:   storage/connect/array.cpp   modified:   storage/connect/array.h   modified:   storage/connect/blkfil.cpp   modified:   storage/connect/blkfil.h   modified:   storage/connect/block.h   modified:   storage/connect/colblk.cpp   modified:   storage/connect/colblk.h   modified:   storage/connect/csort.h   modified:   storage/connect/filamvct.cpp   modified:   storage/connect/filter.cpp   modified:   storage/connect/filter.h   modified:   storage/connect/global.h   modified:   storage/connect/json.h   modified:   storage/connect/plgdbsem.h   modified:   storage/connect/plgdbutl.cpp   modified:   storage/connect/tabcol.cpp   modified:   storage/connect/tabcol.h   modified:   storage/connect/tabdos.cpp   modified:   storage/connect/tabdos.h   modified:   storage/connect/tabjson.cpp   modified:   storage/connect/table.cpp   modified:   storage/connect/tabodbc.cpp   modified:   storage/connect/tabodbc.h   modified:   storage/connect/tabsys.h   modified:   storage/connect/tabxml.h   modified:   storage/connect/value.cpp   modified:   storage/connect/value.h   modified:   storage/connect/xindex.cpp   modified:   storage/connect/xindex.h   modified:   storage/connect/xobject.cpp   modified:   storage/connect/xobject.h   modified:   storage/connect/xtable.h
Set values as nullable when retrieving catalog info
  modified:   storage/connect/jdbconn.cpp
  modified:   storage/connect/mysql-test/connect/r/odbc_oracle.result
  modified:   storage/connect/odbconn.cpp

Change format of Jpath
  modified:   storage/connect/json.cpp
  modified:   storage/connect/jsonudf.cpp
  modified:   storage/connect/mysql-test/connect/r/json.result
  modified:   storage/connect/mysql-test/connect/r/json_udf.result
  modified:   storage/connect/mysql-test/connect/r/json_udf_bin.result
  modified:   storage/connect/mysql-test/connect/r/zip.result
  modified:   storage/connect/mysql-test/connect/t/json.test
  modified:   storage/connect/mysql-test/connect/t/json_udf.test
  modified:   storage/connect/mysql-test/connect/t/json_udf_bin.test
  modified:   storage/connect/mysql-test/connect/t/zip.test
  modified:   storage/connect/tabjson.cpp
  modified:   storage/connect/tabjson.h
  modified:   storage/connect/tabmgo.cpp

Change null representation from ??? to <null>
  modified:   storage/connect/json.cpp

Change the name of UDF that are equal to a native JSON function name
  modified:   storage/connect/jsonudf.cpp
  modified:   storage/connect/jsonudf.h
  modified:   storage/connect/mysql-test/connect/t/json_udf.inc
  modified:   storage/connect/mysql-test/connect/t/json_udf2.inc

Fix bug in making JSON project info
  modified:   storage/connect/mongofam.cpp

Fix COMPUTE when one argument is null
  modified:   storage/connect/value.cpp

Value is null only when nullable
  modified:   storage/connect/value.h

+++ storage/connect/ha_connect.cc
@@ -1045,8 +1045,8 @@ PCSZ GetListOption(PGLOBAL g, PCSZ opname, PCSZ oplist, PCSZ def)
     pv= strchr(pk, '=');
 
     if (pv && (!pn || pv < pn)) {
-			n= MY_MIN(pv - pk, sizeof(key) - 1);
-      memcpy(key, pk, n);
+			n = MY_MIN(static_cast<size_t>(pv - pk), sizeof(key) - 1);
+			memcpy(key, pk, n);
       key[n]= 0;
       pv++;
 			n= MY_MIN((pn ? pn - pv : strlen(pv)), sizeof(val) - 1);
